change ingredients model
